Fixed libmquark to use uintptr_t. Fixes about half the grammar.y warnings on 64 bit.

+++ shotgun/external_libs/libmquark/quark.h
@@ -1,6 +1,7 @@
 #ifndef __MASS_QUARK__
 #define __MASS_QUARK__
-typedef unsigned int quark;
+#include <stdint.h>
+typedef uintptr_t quark;
 
 quark quark_from_string(const char *string);
